Fix table of contents and simplify package catalog
This merge adds `index.md` files in each subdirectory of the docs
to help ensure that the table of contents is laid out intuitively.

The table providing the package catalog has been updated so that
different versions of dependencies for different python versions
are split into multiple rows for readability.