Update to v1.12.0rc2
Changes default version of python to 3.13 and release versions to
be 3.11, 3.12, and 3.13. To support the new python versions all
CDAT related packages have been removed from the `meta.yaml`
and the associated testing of `cdms2` in the `bootstrap.py` file.
The `pcmdi_metrics` package is currently commented out of the
`meta.yaml` because it has `genutils`, `cdutils`, `cdms2`, and `cdp`
as dependencies. When a new version of `pcmdi_metrics`, without CDAT
dependencies, is released it will be uncommented out of the `meta.yaml`.

With the update to python 3.13 the build string for the python version
has changed and is now dealt with accordingly.