Shorten e3sm-unified deployment path
This is necessary to avoid a SbangPathError in spack on Frontier.

This merge also drops the requirement that the machine be specified.
There is no longer an attempt to use a shared conda base environment
as the conda environment for bootstrapping.  Instead, the person
deploying Unified must supply the `--conda` flag or have a base
environment loaded before deploying.